feat: refactor main application structure and enhance cookie management
- Introduced separate handlers for API, cookies, language, and store management.
- Implemented persistent session handling for cookies.
- Added constants for cookie configuration and authentication.
- Enhanced auto-updater setup with forced updates for development.
- Created a new API client with authentication token management.
- Improved window management and IPC handlers for better application control.
