+++ electron/main/constants.ts
@@ -1,0 +1,8 @@
+// Cookie configuration constants
+export const COOKIE_DOMAIN = 'localhost' // Change to your domain in production
+export const AUTH_COOKIE_NAME = 'better-auth.session_token'
+export const USER_DATA_COOKIE_NAME = 'better-auth.session_data'
+
+// Store keys for authentication backup
+export const AUTH_STORE_TOKEN_KEY = 'auth.token'
+export const AUTH_STORE_USER_KEY = 'auth.user'
